+++ Source/Engine/Graphics/Materials/DeformableMaterialShader.cpp
@@ -100,7 +100,7 @@ void DeformableMaterialShader::Bind(BindParameters& params)
     }
 
     // Select pipeline state based on current pass and render mode
-    const bool wireframe = (_info.FeaturesFlags & MaterialFeaturesFlags::Wireframe) != 0 || view.Mode == ViewMode::Wireframe;
+    const bool wireframe = static_cast<int32>(_info.FeaturesFlags & MaterialFeaturesFlags::Wireframe) != 0 || view.Mode == ViewMode::Wireframe;
     CullMode cullMode = view.Pass == DrawPass::Depth ? CullMode::TwoSided : _info.CullMode;
     if (cullMode != CullMode::TwoSided && drawCall.WorldDeterminantSign < 0)
     {
@@ -130,8 +130,8 @@ bool DeformableMaterialShader::Load()
 {
     _drawModes = DrawPass::Depth | DrawPass::QuadOverdraw;
     auto psDesc = GPUPipelineState::Description::Default;
-    psDesc.DepthTestEnable = (_info.FeaturesFlags & MaterialFeaturesFlags::DisableDepthTest) == 0;
-    psDesc.DepthWriteEnable = (_info.FeaturesFlags & MaterialFeaturesFlags::DisableDepthWrite) == 0;
+    psDesc.DepthTestEnable = static_cast<int32>(_info.FeaturesFlags & MaterialFeaturesFlags::DisableDepthTest) == 0;
+    psDesc.DepthWriteEnable = static_cast<int32>(_info.FeaturesFlags & MaterialFeaturesFlags::DisableDepthWrite) == 0;
 
     // Check if use tessellation (both material and runtime supports it)
     const bool useTess = _info.TessellationMode != TessellationMethod::None && GPUDevice::Instance->Limits.HasTessellation;
