+++ Source/Engine/Graphics/Materials/MaterialShaderFeatures.cpp
@@ -130,7 +130,7 @@ bool LightmapFeature::Bind(MaterialShader::BindParameters& params, Span<byte>& c
     auto& data = *(Data*)cb.Get();
     ASSERT_LOW_LAYER(cb.Length() >= sizeof(Data));
 
-    const bool useLightmap = view.Flags & ViewFlags::GI
+    const bool useLightmap = static_cast<int32>(view.Flags & ViewFlags::GI)
 #if USE_EDITOR
             && EnableLightmapsUsage
 #endif
@@ -159,7 +159,7 @@ bool GlobalIlluminationFeature::Bind(MaterialShader::BindParameters& params, Spa
     ASSERT_LOW_LAYER(cb.Length() >= sizeof(Data));
 
     bool useGI = false;
-    if (params.RenderContext.View.Flags & ViewFlags::GI)
+    if (static_cast<int32>(params.RenderContext.View.Flags & ViewFlags::GI))
     {
         switch (params.RenderContext.List->Settings.GlobalIllumination.Mode)
         {
